==> src/eda/03_model_overview.Rmd <==
---
title: "BC Chronic Disease - Overview of Temporal Models Fit with INLA"
date: "2022-05-31"
output:
  html_document:
    df_print: paged
    code_folding: hide
---

### Summary

In our current model, we used an INLA model with a Gamma distribution 
for the likelihood, which was unable to fit data with zeroes.
The purpose of this analysis is to tabulate the models that were fit 
and could not be fit for each rate, disease, and CHSA combination
using this modeling approach.

Overall, 84% of incidence rate data, 91% of HSC prevalence rate data, and
98% of life prevalence rate data were able to be fit with this modeling approach.

```{r message=FALSE}
library(tidyverse)
library(here)
library(knitr)
```

```{r Load Data, message=FALSE}
path <- here("data", "processed")

inc_rate_df <- read_csv(here(path, "incidence_rate_combined.csv"),
  col_types = cols(CLNT_GENDER_LABEL = col_factor())
)
hsc_prev_df <- read_csv(here(path, "hsc_prevalence_combined.csv"),
  col_types = cols(CLNT_GENDER_LABEL = col_factor())
)
life_prev_df <- read_csv(here(path, "life_prevalence_combined.csv"),
  col_types = cols(CLNT_GENDER_LABEL = col_factor())
)
```

```{r Tabulate Function, message=FALSE}
tabulate_models_fit <- function(data) {
  data |>
    group_by(DISEASE, HEALTH_BOUNDARIES) |>
    summarise(MODEL_FIT = case_when(
      anyNA(STD_RATE_PER_1000) ~ NA_real_,
      any(STD_RATE_PER_1000 == 0) ~ 0,
      TRUE ~ 1
    ))
}

inc_models <- tabulate_models_fit(inc_rate_df)
hsc_models <- tabulate_models_fit(hsc_prev_df)
life_models <- tabulate_models_fit(life_prev_df)
```

# Standardized Incidence Rate

Standardized incidence rate models were fit on `r round(4122 / (4122 + 778) * 100, 0)`% of the data.

```{r}
kable(inc_models |>
  ungroup() |>
  count(MODEL_FIT),
  caption = "Table 1. Summary of Age-Standardized Incidence Rate Models."
)
```
In the column `MODEL_FIT`, 1 = Model was fit, 
0 = Model was not fit, and NA = CHSA data was suppressed.


The most frequently affected diseases for Incidence Rate 
were Juvenile Arthritis, MS, and Haemorrhagic Stroke.

```{r}
kable(inc_models |>
  filter(MODEL_FIT == 0) |>
  group_by(DISEASE) |>
  count(MODEL_FIT) |>
  arrange(desc(n)),
  caption = "Table 2. Models Not Fit on Age-Standardized Incidence Rate per Disease."
)
```

# Standardized HSC Prevalence Rate

Standardized HSC prevalence models were fit on `r round(1966 / (1966 + 190) * 100, 0)`% of the data.

```{r}
kable(hsc_models |>
  ungroup() |>
  count(MODEL_FIT),
  caption = "Table 3. Summary of Age-Standardized HSC Prevalence Rate Models."
)
```

The most frequently affected diseases for HSC Prevalence
were forms of stroke (Haemorrhagic, TIA, ischemic).

```{r}
kable(hsc_models |>
  filter(MODEL_FIT == 0) |>
  group_by(DISEASE) |>
  count(MODEL_FIT) |>
  arrange(desc(n)),
  caption = "Table 4. Models Not Fit on Age-Standardized HSC Prevalence Rate per Disease."
)
```

# Standardized Life Prevalence Rate

Standardized life prevalence models were fit on `r round(4812 / (4812 + 88) * 100, 0)`% of the data.

```{r}
kable(
  life_models |>
    ungroup() |>
    count(MODEL_FIT),
  caption = "Table 5. Summary of Age-Standardized Life Prevalence Rate Models."
)
```

The most frequently affected disease for Life Prevalence 
was Juvenile Arthritis.

```{r}
kable(
life_models |>
  filter(MODEL_FIT == 0) |>
  group_by(DISEASE) |>
  count(MODEL_FIT) |>
  arrange(desc(n)),
  caption = "Table 6. Models Not Fit on Age-Standardized Life Prevalence Rate per Disease."
)
```
