==> src/eda/05_modeling_eda_gamma_2.Rmd <==
---
title: 'BC Chronic Disease: Temporal Modeling with INLA (Gamma with modified zeroes)'
author: "Jennifer Hoang"
date: "June 6, 2022"
output:
  html_document:
    df_print: paged
    code_folding: hide
---

```{r message=FALSE}
library(tidyverse)
library(ggplot2)
library(INLA)
library(here)
```

```{r Load Data, message=FALSE}
path <- here("data", "processed")

inc_rate_df <- read_csv(here(path, "incidence_rate_combined.csv"), col_types = cols(CLNT_GENDER_LABEL = col_factor()))
# hsc_prev_df <- read_csv(here(path, "hsc_prevalence_combined.csv"), col_types = cols(CLNT_GENDER_LABEL = col_factor()))
# life_prev_df <- read_csv(here(path, "life_prevalence_combined.csv"), col_types = cols(CLNT_GENDER_LABEL = col_factor()))

x_year <- unique(inc_rate_df$YEAR)
```

## Gamma Distribution with Modified Zeroes

Previously, the Gamma distribution resulted in errors for data from CHSA/diseases with zeroes.
We'll try replacing the zeroes will a very small value, 
such as a value that is 100 times smaller than the the smallest non-zero observation in the dataset.

```{r}
inc_rate_df |>
  filter(!is.na(STD_RATE_PER_1000) & STD_RATE_PER_1000 != 0) |>
  arrange(STD_RATE_PER_1000) |>
  select(DISEASE, STD_RATE_PER_1000) |>
  head(5)
```

The smallest standardized incidence rate is 0.0172,
so we'll try replacing zeroes with 0.0001 in the incidence dataset.

```{r}
inc_rate_processed <- inc_rate_df |>
  mutate(STD_RATE_PER_1000_2 = if_else(STD_RATE_PER_1000 == 0, 0.0001, STD_RATE_PER_1000 ))
```


```{r Functions to Fit and Plot Models}
#' Fit a random walk 1 model with STD_RATE_PER_1000 as response variable and 
#' YEAR as explanatory variable with INLA. Assumes Gamma likelihood.
#'
#' @param data data frame containing original data
#' @return data frame with model results
fit_inla_rw1 <- function(data) {
  model <- inla(STD_RATE_PER_1000_2 ~ 1 + f(YEAR, model = "rw1"),
    data = data,
    family = "gamma",
    control.compute = list(dic = TRUE, waic = TRUE)
  )

  model_results <- tibble(
    DISEASE = data$DISEASE,
    HEALTH_BOUNDARIES = data$HEALTH_BOUNDARIES, 
    x = x_year,
    y_obs = data$STD_RATE_PER_1000,
    y_fitted = model$summary.fitted$mean,
    ci_95_ll = model$summary.fitted$`0.025quant`,
    ci_95_ul = model$summary.fitted$`0.975quant`,
    dic = model$dic$dic,
    waic = model$waic$waic,
    model = "rw1"
  )
  
  model_results
}

#' Fit a random walk 2 model with STD_RATE_PER_1000 as response variable and 
#' YEAR as explanatory variable with INLA. Assumes Gamma likelihood.
#'
#' @param data data frame containing original data
#' @return data frame with model results
fit_inla_rw2 <- function(data) {
  model <- inla(STD_RATE_PER_1000_2 ~ 1 + f(YEAR, model = "rw2"),
    data = data,
    family = "gamma",
    control.compute = list(dic = TRUE, waic = TRUE)
  )

  model_results <- tibble(
    DISEASE = data$DISEASE,
    HEALTH_BOUNDARIES = data$HEALTH_BOUNDARIES, 
    x = x_year,
    y_obs = data$STD_RATE_PER_1000,
    y_fitted = model$summary.fitted$mean,
    ci_95_ll = model$summary.fitted$`0.025quant`,
    ci_95_ul = model$summary.fitted$`0.975quant`,
    dic = model$dic$dic,
    waic = model$waic$waic,
    model = "rw2"
  )
  
  model_results
}

#' Fit a loess model with STD_RATE_PER_1000 as response variable and 
#' YEAR as explanatory variable. Confidence intervals generated
#' using t-distribution approximation
#'
#' @param data data frame containing original data
#' @return data frame with model results
fit_loess <- function(data) {
  model <- loess(STD_RATE_PER_1000_2 ~ YEAR, data = data)
  model_predict <- predict(model, data, se = TRUE)

  t <- qt(p = 0.975, df = model_predict$df)

  model_results <- tibble(
    DISEASE = data$DISEASE[1],
    HEALTH_BOUNDARIES = data$HEALTH_BOUNDARIES[1],
    x = x_year,
    y_obs = data$STD_RATE_PER_1000,
    y_fitted = model_predict$fit,
    ci_95_ll = model_predict$fit - t * model_predict$se.fit,
    ci_95_ul = model_predict$fit + t * model_predict$se.fit,    
    dic = NA,
    waic = NA,
    model = "loess",
  )
  
  model_results
}

#' Generate line plot to compare models and confidence intervals
#'
#' @param data data frame containing original data
#' @param model_df data frame containing model results
#' @param disease character corresponding to entry in DISEASE column
#' @param chsa character corresponding to entry in HEALTH_BOUNDARIES column
#' @param rate_label character (label for Y axis)
#' @return ggplot object
plot_models <- function(data, model_df, disease, chsa, rate_label) {
  cbPalette <- c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7")
  
  ggplot(aes(x = YEAR, y = STD_RATE_PER_1000_2), data = data) +
    geom_point() +
    geom_ribbon(
      aes(x = x, y = y_fitted, ymin = ci_95_ll, ymax = ci_95_ul, fill = model), 
      alpha = 0.2, data = model_df
    ) +
    geom_line(aes(x = x, y = y_fitted, color = model), data = model_df, linetype = 2) +
    labs(
      x = "Year", y = rate_label,
      title = paste(disease, "in", chsa)
    ) +
    scale_fill_manual(values=cbPalette) +
    scale_colour_manual(values=cbPalette)
}
```

```{r Compare 3 Models for 1 Outcome, fig.show="hold", out.width="25%"}
#' Generate line plot to compare loess, random walk 1, and random walk 2 models
#' Only fits data without 0 values due to limitations with Gamma likelihood.
#'
#' @param data data frame
#' @param chsa character corresponding to entry in HEALTH_BOUNDARIES column
#' @param rate_label character (label for Y axis)
#' @return ggplot object
compare_models <- function(data, chsa, rate_label) {
  diseases <- unique(data$DISEASE)
  all_model_df <- tibble()
  plot_df <- tibble()

  for (d in diseases) {
    disease_data <- data |>
      filter(DISEASE == d & HEALTH_BOUNDARIES == chsa)

    if (any(disease_data$STD_RATE_PER_1000 == 0)) {
      tryCatch(
        {plot_df <- rbind(fit_inla_rw1(disease_data), fit_inla_rw2(disease_data), fit_loess(disease_data))
        
          p <- plot_models(disease_data, plot_df, d, chsa, rate_label)
          
          print(p)
        },
        error = function(e) {
          print(paste("Error with", d, chsa))
        }
      )
    }
    all_model_df <- rbind(all_model_df, plot_df)
  }
  all_model_df
}
```


```{r Wrangle Model Criteria}
wrangle_model_results <- function(data) {
  data |>
  select(DISEASE, HEALTH_BOUNDARIES, model, dic, waic) |>
  filter(model != "loess") |>
  unique() |>
  pivot_wider(names_from = model,
              values_from = c(dic, waic)) |>
  mutate(best_model_dic = ifelse(dic_rw1 < dic_rw2, "rw1", "rw2"),
        best_model_waic = ifelse(waic_rw1 < waic_rw2, "rw1", "rw2")) |>
  mutate_if(is.numeric, round, 2) 
}
```


### Age-Standardized Incidence Rate in Vanderhoof Rural

This region has a lot of diseases with zero values. 
Let's see how the models perform on this region.

We'll only run the model on the diseases with at least one zero value from this CHSA.

```{r fig.show="hold", message=FALSE, warning=FALSE, out.width="50%"}
inc_vr <- compare_models(inc_rate_processed, "5232 Vanderhoof Rural", "Age-Std Incidence Rate")

wrangle_model_results(inc_vr)
```

Many of the rw2 models display very large credible intervals, even CKD and MS, where the rw2 was chosen.
The rw2 CI ranges are much wider than our baseline loess model.

Perhaps we can also exclude the data with only zeroes, 
such as Juvenile Arthritis, since these models don't look very useful.

Let's try this again on Vanderhoof but only with rw1 models.


```{r}
#' Generate line plot to compare loess, random walk 1, and random walk 2 models
#' Only fits data without 0 values due to limitations with Gamma likelihood.
#'
#' @param data data frame
#' @param chsa character corresponding to entry in HEALTH_BOUNDARIES column
#' @param rate_label character (label for Y axis)
#' @return ggplot object
compare_models_rw1 <- function(data, chsa, rate_label) {
  diseases <- unique(data$DISEASE)
  all_model_df <- tibble()
  plot_df <- tibble()

  for (d in diseases) {
    disease_data <- data |>
      filter(DISEASE == d & HEALTH_BOUNDARIES == chsa)

    if (any(disease_data$STD_RATE_PER_1000 == 0)) {
      tryCatch(
        {plot_df <- rbind(fit_inla_rw1(disease_data), fit_loess(disease_data))
        
          p <- plot_models(disease_data, plot_df, d, chsa, rate_label)
          
          print(p)
        },
        error = function(e) {
          print(paste("Error with", d, chsa))
        }
      )
    }
    all_model_df <- rbind(all_model_df, plot_df)
  }
  all_model_df
}

wrangle_model_results_rw1 <- function(data) {
  data |>
  select(DISEASE, HEALTH_BOUNDARIES, model, dic, waic) |>
  filter(model != "loess") |>
  unique() |>
  pivot_wider(names_from = model,
              values_from = c(dic, waic)) |>
  mutate_if(is.numeric, round, 2) 
}
```

### Vanderhoof Rural RW1 vs. LOESS

```{r fig.show="hold", message=FALSE, warning=FALSE, out.width="50%"}
inc_vr_2 <- compare_models_rw1(inc_rate_processed, "5232 Vanderhoof Rural", "Age-Std Incidence Rate")

wrangle_model_results_rw1(inc_vr_2)
```

The RW1 model seems to either underfit or overfit the data in some cases.

## Part 2: Vanderhoof Rural RW1 with Penalized Complexity Prior

Let's try this again with a different prior.

The Random Walk has a precision parameter that is defined on theta. (https://inla.r-inla-download.org/r-inla.org/doc/latent/rw1.pdf)
Lower values of theta result in more smoothing. (https://www.seascapemodels.org/rstats/2017/06/21/bayesian-smoothing.html)
The default prior for theta is set using loggamma with a value of 5e-05 (vague prior), 
but another recommended prior is the "pc.prec" or Penalized Complexity prior with the setting "param=c(1,0.01)".

It's also recommended to set the "scale.model = TRUE" for random walk models (https://becarioprecario.bitbucket.io/inla-gitbook/ch-priors.html).


```{r}
#' Fit a random walk 1 model with STD_RATE_PER_1000 as response variable and
#' YEAR as explanatory variable with INLA. Assumes Gamma likelihood.
#' Prior set using Penalized Complexity. Scaled model.
#'
#' @param data data frame containing original data
#' @return data frame with model results
fit_inla_rw1 <- function(data) {
  model <- inla(STD_RATE_PER_1000_2 ~ 1 + f(YEAR,
    model = "rw1",
    scale.model = TRUE,
    hyper = list(theta = list(prior = "pc.prec", param = c(1, 0.01)))
  ),
  data = data,
  family = "gamma",
  control.compute = list(dic = TRUE, waic = TRUE)
  )

  model_results <- tibble(
    DISEASE = data$DISEASE,
    HEALTH_BOUNDARIES = data$HEALTH_BOUNDARIES,
    x = x_year,
    y_obs = data$STD_RATE_PER_1000,
    y_fitted = model$summary.fitted$mean,
    ci_95_ll = model$summary.fitted$`0.025quant`,
    ci_95_ul = model$summary.fitted$`0.975quant`,
    dic = model$dic$dic,
    waic = model$waic$waic,
    model = "rw1"
  )

  model_results
}

#' Fit a random walk 2 model with STD_RATE_PER_1000 as response variable and
#' YEAR as explanatory variable with INLA. Assumes Gamma likelihood.
#' Prior set using Penalized Complexity. Scaled model.
#'
#' @param data data frame containing original data
#' @return data frame with model results
fit_inla_rw2 <- function(data) {
  model <- inla(STD_RATE_PER_1000_2 ~ 1 + f(YEAR,
    model = "rw2",
    scale.model = TRUE,
    hyper = list(theta = list(prior = "pc.prec", param = c(1, 0.01)))
  ),
  data = data,
  family = "gamma",
  control.compute = list(dic = TRUE, waic = TRUE)
  )

  model_results <- tibble(
    DISEASE = data$DISEASE,
    HEALTH_BOUNDARIES = data$HEALTH_BOUNDARIES,
    x = x_year,
    y_obs = data$STD_RATE_PER_1000,
    y_fitted = model$summary.fitted$mean,
    ci_95_ll = model$summary.fitted$`0.025quant`,
    ci_95_ul = model$summary.fitted$`0.975quant`,
    dic = model$dic$dic,
    waic = model$waic$waic,
    model = "rw2"
  )

  model_results
}
```


```{r fig.show="hold", message=FALSE, warning=FALSE, out.width="50%"}
inc_vr_3 <- compare_models(inc_rate_processed, "5232 Vanderhoof Rural", "Age-Std Incidence Rate")

wrangle_model_results(inc_vr_3)
```

The RW2 model still produces large intervals after adjusting the prior. Let's plot only the rw1 and LOESS models

### Vanderhoof Rural RW1 vs. LOESS (PC Prior)

```{r fig.show="hold", message=FALSE, warning=FALSE, out.width="50%"}
inc_vr_4 <- compare_models_rw1(inc_rate_processed, "5232 Vanderhoof Rural", "Age-Std Incidence Rate")

wrangle_model_results_rw1(inc_vr_4)
```

The RW1 with PC prior and scaled model looks to be an improvement over the original RW1 prior.
